# Change implementation of BlobSet