--- a/internal/repository/repack.go
+++ b/internal/repository/repack.go
@@ -25,7 +25,7 @@ const numRepackWorkers = 8
 // The map keepBlobs is modified by Repack, it is used to keep track of which
 // blobs have been processed.
 func Repack(ctx context.Context, repo restic.Repository, packs restic.IDSet, keepBlobs restic.BlobSet, p *progress.Counter) (obsoletePacks restic.IDSet, err error) {
-	debug.Log("repacking %d packs while keeping %d blobs", len(packs), len(keepBlobs))
+	debug.Log("repacking %d packs while keeping %d blobs", len(packs), keepBlobs.Len())
 
 	wg, wgCtx := errgroup.WithContext(ctx)