--- a/internal/restic/blob_set.go
+++ b/internal/restic/blob_set.go
@@ -3,42 +3,65 @@ package restic
 import "sort"
 
 // BlobSet is a set of blobs.
-type BlobSet map[BlobHandle]struct{}
+type BlobSet struct {
+	byType [NumBlobTypes]IDSet
+}
 
-// NewBlobSet returns a new BlobSet, populated with ids.
+// NewBlobSet returns a new blobSet, populated with blob handles.
+// Note that this is not thread-safe, but writing to blobs of different types
+// concurrently is safe.
 func NewBlobSet(handles ...BlobHandle) BlobSet {
-	m := make(BlobSet)
+	var m BlobSet
+	for t := InvalidBlob; t < NumBlobTypes; t++ {
+		m.byType[t] = NewIDSet()
+	}
 	for _, h := range handles {
-		m[h] = struct{}{}
+		m.byType[h.Type].Insert(h.ID)
 	}
 
 	return m
 }
 
+// ForAll calls the given func for all entries in the blobSet
+func (s BlobSet) ForAll(fn func(h BlobHandle) error) error {
+	for t := InvalidBlob; t < NumBlobTypes; t++ {
+		for id := range s.byType[t] {
+			err := fn(BlobHandle{Type: t, ID: id})
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+// Len returns the number of elements in the blobSet
+func (s BlobSet) Len() (length int) {
+	for t := InvalidBlob; t < NumBlobTypes; t++ {
+		length += len(s.byType[t])
+	}
+	return length
+}
+
 // Has returns true iff id is contained in the set.
 func (s BlobSet) Has(h BlobHandle) bool {
-	_, ok := s[h]
-	return ok
+	return s.byType[h.Type].Has(h.ID)
 }
 
 // Insert adds id to the set.
 func (s BlobSet) Insert(h BlobHandle) {
-	s[h] = struct{}{}
+	s.byType[h.Type].Insert(h.ID)
 }
 
 // Delete removes id from the set.
 func (s BlobSet) Delete(h BlobHandle) {
-	delete(s, h)
+	delete(s.byType[h.Type], h.ID)
 }
 
 // Equals returns true iff s equals other.
 func (s BlobSet) Equals(other BlobSet) bool {
-	if len(s) != len(other) {
-		return false
-	}
-
-	for h := range s {
-		if _, ok := other[h]; !ok {
+	for t := InvalidBlob; t < NumBlobTypes; t++ {
+		if !s.byType[t].Equals(other.byType[t]) {
 			return false
 		}
 	}
@@ -48,26 +71,30 @@ func (s BlobSet) Equals(other BlobSet) bool {
 
 // Merge adds the blobs in other to the current set.
 func (s BlobSet) Merge(other BlobSet) {
-	for h := range other {
-		s.Insert(h)
+	for t := InvalidBlob; t < NumBlobTypes; t++ {
+		for id := range other.byType[t] {
+			s.byType[t].Insert(id)
+		}
 	}
 }
 
 // Intersect returns a new set containing the handles that are present in both sets.
-func (s BlobSet) Intersect(other BlobSet) (result BlobSet) {
-	result = NewBlobSet()
+func (s BlobSet) Intersect(other BlobSet) BlobSet {
+	result := NewBlobSet()
 
 	set1 := s
 	set2 := other
 
-	// iterate over the smaller set
-	if len(set2) < len(set1) {
-		set1, set2 = set2, set1
-	}
+	for t := InvalidBlob; t < NumBlobTypes; t++ {
+		// iterate over the smaller set
+		if len(set2.byType[t]) < len(set1.byType[t]) {
+			set1, set2 = set2, set1
+		}
 
-	for h := range set1 {
-		if set2.Has(h) {
-			result.Insert(h)
+		for id := range set1.byType[t] {
+			if set2.byType[t].Has(id) {
+				result.byType[t].Insert(id)
+			}
 		}
 	}
 
@@ -76,11 +103,13 @@ func (s BlobSet) Intersect(other BlobSet) (result BlobSet) {
 
 // Sub returns a new set containing all handles that are present in s but not in
 // other.
-func (s BlobSet) Sub(other BlobSet) (result BlobSet) {
-	result = NewBlobSet()
-	for h := range s {
-		if !other.Has(h) {
-			result.Insert(h)
+func (s BlobSet) Sub(other BlobSet) BlobSet {
+	result := NewBlobSet()
+	for t := InvalidBlob; t < NumBlobTypes; t++ {
+		for id := range s.byType[t] {
+			if !other.byType[t].Has(id) {
+				result.byType[t].Insert(id)
+			}
 		}
 	}
 
@@ -89,9 +118,11 @@ func (s BlobSet) Sub(other BlobSet) (result BlobSet) {
 
 // List returns a sorted slice of all BlobHandle in the set.
 func (s BlobSet) List() BlobHandles {
-	list := make(BlobHandles, 0, len(s))
-	for h := range s {
-		list = append(list, h)
+	list := make(BlobHandles, 0, s.Len())
+	for t := InvalidBlob; t < NumBlobTypes; t++ {
+		for id := range s.byType[t] {
+			list = append(list, BlobHandle{Type: t, ID: id})
+		}
 	}
 
 	sort.Sort(list)