What does this PR change? What problem does it solve?
-----------------------------------------------------

Changes the implementation of `BlobSet` from `map[BlobHandle]struct{}` to `[NumBlobTypes]IDSet`.
This has the following advantages:

- uses less memory for large number of blobs within a BlobSet
- is thread-safe if blobs of different types are added or deleted concurrently (can e.g. be used in #3106)

Was the change discussed in an issue or in the forum before?
------------------------------------------------------------

No. 

Checklist
---------

- [x] I have read the [Contribution Guidelines](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#providing-patches)
- [x] I have enabled [maintainer edits for this PR](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork)
-  I have not added tests for all changes in this PR, but modified the current ones
-  I have not added documentation for the changes (in the manual) - not needed
-  There's no new file in `changelog/unreleased/` that describes the changes for our users (template [here](https://github.com/restic/restic/blob/master/changelog/TEMPLATE)) - user should hardly notice this change
- [x] I have run `gofmt` on the code in all commits
- [x] All commit messages are formatted in the same style as [the other commits in the repo](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#git-commits)
- [x] I'm done, this Pull Request is ready for review