@greatroar 2020-12-17
I'm curious to see if this really uses less memory. I'd implemented [something very similar](greatroar/restic@35c4f8fc84b975ef56acd6269995e9ba92a6d969) earlier, but I found that it actually took slightly more memory for a restic check on a 450GiB repo: 1.722 GiB versus 1.683 GiB max resident.