@aawsome 2020-12-17
> I'm curious to see if this really uses less memory.

I think this is the same discussion we had about the index implementation. The go `map` implementation shows some undesired worst-case behavior especially directly after growing.

Of course as we use two maps instead of one, there might be cases where we now hit the worst-case scenario which we didn't before. But there are other cases where the worst case scenario is hit with one map and this PR would significantly reduce the memory. In summary, we store 32 instead of 33 bytes in two maps instead of one map.

If we are not satisfied with this reduction (or the worst case behavior of go maps), we need to implement something similar as you did with the index (would be nice if go had generics here...).
This was however not my intention with this PR :wink: