@MichaelEischer 2021-07-23
I've hacked together as short benchmark
```go
package restic

import (
	"strconv"
	"testing"
)

func BenchmarkBlobSet(b *testing.B) {
	for k := 0; k < 2_000_000; k += 100_000 {
		name := strconv.FormatInt(int64(k), 10)
		b.Run(name, func(b *testing.B) {
			b.ReportAllocs()
			for c := 0; c < b.N; c++ {
				bs := NewBlobSet()
				for i := 0; i < k; i++ {
					h := NewRandomBlobHandle()
					bs.Insert(h)
				}
			}
		})
	}
}
```

```
name                old time/op    new time/op    delta
BlobSet/0-12          25.7ns ± 1%    80.2ns ± 2%  +211.49%  (p=0.000 n=8+9)
BlobSet/100000-12     94.5ms ±18%   116.6ms ±25%   +23.45%  (p=0.003 n=10+10)
BlobSet/200000-12      189ms ±22%     229ms ±23%   +21.29%  (p=0.009 n=10+10)
BlobSet/300000-12      284ms ±32%     343ms ±18%   +20.66%  (p=0.015 n=10+10)
BlobSet/400000-12      372ms ±24%     482ms ±33%   +29.75%  (p=0.007 n=10+10)
BlobSet/500000-12      511ms ±21%     624ms ±31%      ~     (p=0.052 n=10+10)
BlobSet/600000-12      566ms ±11%     691ms ±18%   +22.02%  (p=0.000 n=9+9)
BlobSet/700000-12      705ms ±24%     735ms ±57%      ~     (p=0.684 n=10+10)
BlobSet/800000-12      668ms ±25%     917ms ±37%   +37.22%  (p=0.009 n=10+10)
BlobSet/900000-12      838ms ±28%    1038ms ±23%   +23.96%  (p=0.004 n=10+9)
BlobSet/1000000-12     954ms ±26%    1130ms ±30%   +18.42%  (p=0.043 n=10+10)
BlobSet/1100000-12     1.10s ±42%     1.23s ±18%      ~     (p=0.247 n=10+10)
BlobSet/1200000-12     1.19s ±20%     1.38s ±31%      ~     (p=0.063 n=10+10)
BlobSet/1300000-12     1.24s ±36%     1.51s ±23%   +21.16%  (p=0.019 n=10+10)
BlobSet/1400000-12     1.26s ±25%     1.63s ±38%   +29.50%  (p=0.005 n=10+10)
BlobSet/1500000-12     1.33s ±16%     1.46s ±34%      ~     (p=0.796 n=10+10)
BlobSet/1600000-12     1.51s ±26%     1.91s ±30%   +27.06%  (p=0.004 n=10+9)
BlobSet/1700000-12     1.45s ±10%     1.78s ±35%   +22.21%  (p=0.019 n=7+10)
BlobSet/1800000-12     1.71s ±32%     2.02s ±33%   +18.14%  (p=0.029 n=10+10)
BlobSet/1900000-12     1.76s ±38%     2.32s ±27%   +31.71%  (p=0.003 n=10+10)

name                old alloc/op   new alloc/op   delta
BlobSet/0-12           48.0B ± 0%    144.0B ± 0%  +200.00%  (p=0.000 n=10+10)
BlobSet/100000-12     14.3MB ± 0%    14.0MB ± 0%    -1.97%  (p=0.000 n=10+10)
BlobSet/200000-12     28.6MB ± 0%    28.1MB ± 0%    -1.96%  (p=0.000 n=10+10)
BlobSet/300000-12     51.9MB ± 0%    50.7MB ± 0%    -2.15%  (p=0.000 n=10+10)
BlobSet/400000-12     57.3MB ± 0%    56.2MB ± 0%    -1.95%  (p=0.000 n=10+10)
BlobSet/500000-12      100MB ± 0%      98MB ± 0%    -2.22%  (p=0.000 n=10+10)
BlobSet/600000-12      104MB ± 0%     102MB ± 0%    -2.19%  (p=0.000 n=10+10)
BlobSet/700000-12      108MB ± 0%     106MB ± 0%    -2.04%  (p=0.000 n=10+10)
BlobSet/800000-12      115MB ± 0%     112MB ± 0%    -1.95%  (p=0.000 n=10+10)
BlobSet/900000-12      198MB ± 0%     193MB ± 0%    -2.26%  (p=0.000 n=10+10)
BlobSet/1000000-12     201MB ± 0%     197MB ± 0%    -2.22%  (p=0.000 n=10+9)
BlobSet/1100000-12     205MB ± 0%     200MB ± 0%    -2.17%  (p=0.000 n=10+10)
BlobSet/1200000-12     208MB ± 0%     203MB ± 0%    -2.15%  (p=0.000 n=9+8)
BlobSet/1300000-12     212MB ± 0%     207MB ± 0%    -2.09%  (p=0.000 n=10+9)
BlobSet/1400000-12     217MB ± 0%     213MB ± 0%    -2.05%  (p=0.000 n=10+10)
BlobSet/1500000-12     223MB ± 0%     219MB ± 0%    -1.99%  (p=0.000 n=10+10)
BlobSet/1600000-12     230MB ± 0%     225MB ± 0%    -1.92%  (p=0.000 n=10+10)
BlobSet/1700000-12     237MB ± 0%     232MB ± 0%    -1.88%  (p=0.000 n=10+10)
BlobSet/1800000-12     396MB ± 0%     387MB ± 0%    -2.25%  (p=0.000 n=10+10)
BlobSet/1900000-12     399MB ± 0%     390MB ± 0%    -2.22%  (p=0.000 n=8+9)

name                old allocs/op  new allocs/op  delta
BlobSet/0-12            1.00 ± 0%      3.00 ± 0%  +200.00%  (p=0.000 n=10+10)
BlobSet/100000-12       104k ± 0%      104k ± 0%    -0.04%  (p=0.000 n=10+10)
BlobSet/200000-12       208k ± 0%      208k ± 0%    -0.02%  (p=0.004 n=10+10)
BlobSet/300000-12       310k ± 0%      310k ± 0%    -0.01%  (p=0.016 n=10+10)
BlobSet/400000-12       416k ± 0%      416k ± 0%      ~     (p=0.324 n=10+10)
BlobSet/500000-12       519k ± 0%      519k ± 0%    -0.01%  (p=0.043 n=10+10)
BlobSet/600000-12       619k ± 0%      619k ± 0%    -0.03%  (p=0.000 n=9+10)
BlobSet/700000-12       723k ± 0%      723k ± 0%      ~     (p=0.631 n=10+10)
BlobSet/800000-12       833k ± 0%      833k ± 0%      ~     (p=0.151 n=10+9)
BlobSet/900000-12       938k ± 0%      938k ± 0%      ~     (p=0.075 n=10+10)
BlobSet/1000000-12     1.04M ± 0%     1.04M ± 0%      ~     (p=0.218 n=10+10)
BlobSet/1100000-12     1.14M ± 0%     1.14M ± 0%      ~     (p=0.739 n=10+10)
BlobSet/1200000-12     1.24M ± 0%     1.24M ± 0%      ~     (p=0.387 n=9+9)
BlobSet/1300000-12     1.34M ± 0%     1.34M ± 0%      ~     (p=0.684 n=10+10)
BlobSet/1400000-12     1.45M ± 0%     1.45M ± 0%      ~     (p=0.813 n=10+10)
BlobSet/1500000-12     1.55M ± 0%     1.55M ± 0%      ~     (p=0.725 n=10+10)
BlobSet/1600000-12     1.66M ± 0%     1.67M ± 0%      ~     (p=0.105 n=10+10)
BlobSet/1700000-12     1.78M ± 0%     1.78M ± 0%      ~     (p=0.853 n=10+10)
BlobSet/1800000-12     1.88M ± 0%     1.88M ± 0%      ~     (p=0.853 n=10+10)
BlobSet/1900000-12     1.98M ± 0%     1.98M ± 0%      ~     (p=0.395 n=8+10)
```

In summary the memory savings seem to be rather low, And the performance is apparently quite a bit worse.

My suggestion would be to close this PR as I don't think that it's worth the additional complexity.